Блок для сделок
Создан отдельный блок для сделок, в который была перенесена модель таблицы с историей сделок. Созданы html заглушки для окон с историей, добавлением и редактированием сделок.
