+++ trackingportfolio/portfolio/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from django.contrib.sessions.models import Session
 
 from . models import *
 
@@ -10,7 +9,3 @@ class PortfolioAdmin(admin.ModelAdmin):
 @admin.register(Asset)
 class AssetAdmin(admin.ModelAdmin):
     list_display = ['ticker', 'name', 'count', 'portfolio']
-
-@admin.register(Transaction)
-class TransactionAdmin(admin.ModelAdmin):
-    list_display = ['type', 'portfolio', 'asset', 'value']
